requirements.txt
Pillow>=10.0.0
pillow-avif-plugin>=1.0.0
tkinterdnd2>=0.4.0